[circt-verilog-lsp][NFC] Break out implementation classes
**Summary**
- Refactored `VerilogServerImpl` into smaller, well-structured components for clarity, testability, and maintainability.
- Moved large inline class implementations out of `VerilogServer.cpp` into dedicated compilation units.

**Changes**
- **New modules**
  - `VerilogTextFile`: manages LSP file contents, versioning, and incremental updates.
  - `VerilogDocument`: handles Slang integration, parsing, diagnostics, and indexing setup.
  - `VerilogIndex`: implements symbol/reference tracking using interval maps.
  - `LSPDiagnosticClient`: converts Slang diagnostics to LSP diagnostics.
  - `VerilogServerContext`: stores server-wide options and shared configuration.
- **Refactors**
  - `VerilogServer`: simplified to a coordinator using the pImpl pattern and `VerilogTextFile` instances.
  - Removed large class definitions (Document, Index, DiagnosticClient, etc.) from `VerilogServer.cpp`.
  - Updated `CMakeLists.txt` to compile new files.
- **Headers**
  - Added clear LLVM-style header comments.
  - Reduced includes and replaced them with forward declarations where possible.

**Impact**
- No functional changes; purely structural refactor.
- Improved code organization and readability.
- Enables independent testing and clearer layering between LSP, Slang, and CIRCT logic.
