+++ lib/Tools/circt-verilog-lsp-server/VerilogServerImpl/VerilogServer.h
@@ -6,8 +6,12 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// Main entry function for circt-verilog-lsp-server for when built as standalone
-// binary.
+// VerilogServer.h
+//
+// This header declares VerilogServer, the top-level coordinator of the CIRCT
+// Verilog LSP server. It manages all open files and acts as the entry point for
+// language server operations such as “didOpen”, “didChange”, “didClose”,
+// “definition”, and “references”.
 //
 //===----------------------------------------------------------------------===//
 
@@ -79,4 +83,4 @@ private:
 } // namespace lsp
 } // namespace circt
 
-#endif // LIB_CIRCT_TOOLS_CIRCT_VERILOG_LSP_SERVER_VERILOGSERVER_H_
+#endif
